matching spam and land cover cropland
Most of the joining process was done in Arc, but this script (which is messy and outdated in other sections) specifically addresses where there are agricultural value cells but no land cover points to join to. This creates centroids of the missing cells and appends to the final agricultural origins file.